+++ sdk/containerregistry/azure-containerregistry/azure/containerregistry/aio/_async_container_registry_client.py
@@ -18,9 +18,10 @@ from azure.core.tracing.decorator import distributed_trace
 from azure.core.tracing.decorator_async import distributed_trace_async
 
 from ._async_base_client import ContainerRegistryBaseClient, AsyncTransportWrapper
-from ._async_container_repository_client import ContainerRepositoryClient
+from ._async_container_repository import ContainerRepository
 from .._generated.models import AcrErrors
 from .._helpers import _parse_next_link
+from ._async_registry_artifact import RegistryArtifact
 from .._models import RepositoryProperties, DeleteRepositoryResult
 
 if TYPE_CHECKING:
@@ -30,7 +31,6 @@ if TYPE_CHECKING:
 class ContainerRegistryClient(ContainerRegistryBaseClient):
     def __init__(self, endpoint: str, credential: "AsyncTokenCredential", **kwargs: Dict[str, Any]) -> None:
         """Create a ContainerRegistryClient from an endpoint and a credential
-
         :param endpoint: An ACR endpoint
         :type endpoint: str
         :param credential: The credential with which to authenticate
@@ -54,11 +54,10 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         super(ContainerRegistryClient, self).__init__(endpoint=endpoint, credential=credential, **kwargs)
 
     @distributed_trace_async
-    async def delete_repository(self, repository: str, **kwargs: Dict[str, Any]) -> DeleteRepositoryResult:
+    async def delete_repository(self, repository_name: str, **kwargs: Dict[str, Any]) -> DeleteRepositoryResult:
         """Delete a repository
 
-        :param repository: The repository to delete
-        :type repository: str
+        :param str repository_name: The repository to delete
         :returns: Object containing information about the deleted repository
         :rtype: :class:`~azure.containerregistry.DeleteRepositoryResult`
         :raises: :class:`~azure.core.exceptions.ResourceNotFoundError`
@@ -72,13 +71,12 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
                 :dedent: 8
                 :caption: Delete a repository from the `ContainerRegistryClient`
         """
-        result = await self._client.container_registry.delete_repository(repository, **kwargs)
+        result = await self._client.container_registry.delete_repository(repository_name, **kwargs)
         return DeleteRepositoryResult._from_generated(result)  # pylint: disable=protected-access
 
     @distributed_trace
-    def list_repositories(self, **kwargs: Dict[str, Any]) -> AsyncItemPaged[str]:
+    def list_repository_names(self, **kwargs: Dict[str, Any]) -> AsyncItemPaged[str]:
         """List all repositories
-
         :keyword last: Query parameter for the last item in the previous call. Ensuing
             call will return values after last lexicallyy
         :paramtype last: str
@@ -93,8 +91,8 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         .. admonition:: Example:
 
             .. literalinclude:: ../samples/async_samples/sample_delete_old_tags_async.py
-                :start-after: [START list_repositories]
-                :end-before: [END list_repositories]
+                :start-after: [START list_repository_names]
+                :end-before: [END list_repository_names]
                 :language: python
                 :dedent: 8
                 :caption: List repositories in a container registry account
@@ -161,7 +159,7 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
             deserialized = self._client._deserialize(  # pylint: disable=protected-access
                 "Repositories", pipeline_response
             )
-            list_of_elem = deserialized.repositories
+            list_of_elem = deserialized.repositories or []
             if cls:
                 list_of_elem = cls(list_of_elem)
             link = None
@@ -189,23 +187,20 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         return AsyncItemPaged(get_next, extract_data)
 
     @distributed_trace
-    def get_repository_client(self, repository: str, **kwargs: Dict[str, Any]) -> ContainerRepositoryClient:
+    def get_repository(self, repository_name: str, **kwargs: Any) -> ContainerRepository:
         """Get a repository client
 
-        :param repository: The repository to create a client for
-        :type repository: str
-        :returns: :class:`~azure.containerregistry.aio.ContainerRepositoryClient`
+        :param str repository_name: The repository to create a client for
+        :returns: :class:`~azure.containerregistry.aio.ContainerRepository`
 
         Example
 
         .. code-block:: python
-
             from azure.containerregistry.aio import ContainerRepositoryClient
             from azure.identity.aio import DefaultAzureCredential
-
             account_url = os.environ["CONTAINERREGISTRY_ENDPOINT"]
             client = ContainerRegistryClient(account_url, DefaultAzureCredential())
-            repository_client = client.get_repository_client("my_repository")
+            repository_client = client.get_repository("my_repository")
         """
         _pipeline = AsyncPipeline(
             transport=AsyncTransportWrapper(
@@ -213,6 +208,28 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
             ),
             policies=self._client._client._pipeline._impl_policies,  # pylint: disable=protected-access
         )
-        return ContainerRepositoryClient(
-            self._endpoint, repository, credential=self._credential, pipeline=_pipeline, **kwargs
+        return ContainerRepository(
+            self._endpoint, repository_name, credential=self._credential, pipeline=_pipeline, **kwargs
+        )
+
+    @distributed_trace
+    def get_artifact(self, repository_name: str, tag_or_digest: str, **kwargs: Dict[str, Any]) -> RegistryArtifact:
+        """Get a Registry Artifact object
+
+        :param str repository_name: Name of the repository
+        :param str tag_or_digest: The tag or digest of the artifact
+        :returns: :class:`~azure.containerregistry.RegistryArtifact`
+        :raises: None
+        """
+        _pipeline = AsyncPipeline(
+            transport=AsyncTransportWrapper(self._client._client._pipeline._transport),  # pylint: disable=protected-access
+            policies=self._client._client._pipeline._impl_policies,  # pylint: disable=protected-access
+        )
+        return RegistryArtifact(
+            self._endpoint,
+            repository_name,
+            tag_or_digest,
+            self._credential,
+            pipeline=_pipeline,
+            **kwargs
         )