+++ sdk/maps/azure-mgmt-maps/azure/mgmt/maps/aio/operations/_accounts_operations.py
@@ -45,7 +45,7 @@ class AccountsOperations:
         self,
         resource_group_name: str,
         account_name: str,
-        maps_account_create_parameters: "_models.MapsAccountCreateParameters",
+        maps_account: "_models.MapsAccount",
         **kwargs
     ) -> "_models.MapsAccount":
         """Create or update a Maps Account. A Maps Account holds the keys which allow access to the Maps
@@ -55,8 +55,8 @@ class AccountsOperations:
         :type resource_group_name: str
         :param account_name: The name of the Maps Account.
         :type account_name: str
-        :param maps_account_create_parameters: The new or updated parameters for the Maps Account.
-        :type maps_account_create_parameters: ~azure.mgmt.maps.models.MapsAccountCreateParameters
+        :param maps_account: The new or updated parameters for the Maps Account.
+        :type maps_account: ~azure.mgmt.maps.models.MapsAccount
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: MapsAccount, or the result of cls(response)
         :rtype: ~azure.mgmt.maps.models.MapsAccount
@@ -67,7 +67,7 @@ class AccountsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-02-01-preview"
+        api_version = "2021-02-01"
         content_type = kwargs.pop("content_type", "application/json")
         accept = "application/json"
 
@@ -90,7 +90,7 @@ class AccountsOperations:
         header_parameters['Accept'] = self._serialize.header("accept", accept, 'str')
 
         body_content_kwargs = {}  # type: Dict[str, Any]
-        body_content = self._serialize.body(maps_account_create_parameters, 'MapsAccountCreateParameters')
+        body_content = self._serialize.body(maps_account, 'MapsAccount')
         body_content_kwargs['content'] = body_content
         request = self._client.put(url, query_parameters, header_parameters, **body_content_kwargs)
         pipeline_response = await self._client._pipeline.run(request, stream=False, **kwargs)
@@ -121,7 +121,7 @@ class AccountsOperations:
         **kwargs
     ) -> "_models.MapsAccount":
         """Updates a Maps Account. Only a subset of the parameters may be updated after creation, such as
-        Sku and Tags.
+        Sku, Tags, Properties.
 
         :param resource_group_name: The name of the resource group. The name is case insensitive.
         :type resource_group_name: str
@@ -139,7 +139,7 @@ class AccountsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-02-01-preview"
+        api_version = "2021-02-01"
         content_type = kwargs.pop("content_type", "application/json")
         accept = "application/json"
 
@@ -203,7 +203,7 @@ class AccountsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-02-01-preview"
+        api_version = "2021-02-01"
         accept = "application/json"
 
         # Construct URL
@@ -259,7 +259,7 @@ class AccountsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-02-01-preview"
+        api_version = "2021-02-01"
         accept = "application/json"
 
         # Construct URL
@@ -315,7 +315,7 @@ class AccountsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-02-01-preview"
+        api_version = "2021-02-01"
         accept = "application/json"
 
         def prepare_request(next_link=None):
@@ -347,7 +347,7 @@ class AccountsOperations:
             list_of_elem = deserialized.value
             if cls:
                 list_of_elem = cls(list_of_elem)
-            return None, AsyncList(list_of_elem)
+            return deserialized.next_link or None, AsyncList(list_of_elem)
 
         async def get_next(next_link=None):
             request = prepare_request(next_link)
@@ -383,7 +383,7 @@ class AccountsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-02-01-preview"
+        api_version = "2021-02-01"
         accept = "application/json"
 
         def prepare_request(next_link=None):
@@ -414,7 +414,7 @@ class AccountsOperations:
             list_of_elem = deserialized.value
             if cls:
                 list_of_elem = cls(list_of_elem)
-            return None, AsyncList(list_of_elem)
+            return deserialized.next_link or None, AsyncList(list_of_elem)
 
         async def get_next(next_link=None):
             request = prepare_request(next_link)
@@ -458,7 +458,7 @@ class AccountsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-02-01-preview"
+        api_version = "2021-02-01"
         accept = "application/json"
 
         # Construct URL
@@ -521,7 +521,7 @@ class AccountsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-02-01-preview"
+        api_version = "2021-02-01"
         content_type = kwargs.pop("content_type", "application/json")
         accept = "application/json"