+++ sdk/resources/azure-mgmt-resource/azure/mgmt/resource/resources/v2018_02_01/models/_models.py
@@ -166,11 +166,11 @@ class DeploymentExportResult(msrest.serialization.Model):
     """The deployment export result.
 
     :param template: The template content.
-    :type template: object
+    :type template: str
     """
 
     _attribute_map = {
-        'template': {'key': 'template', 'type': 'object'},
+        'template': {'key': 'template', 'type': 'str'},
     }
 
     def __init__(
@@ -315,7 +315,7 @@ class DeploymentOperationProperties(msrest.serialization.Model):
     :ivar status_code: Operation status code.
     :vartype status_code: str
     :ivar status_message: Operation status message.
-    :vartype status_message: object
+    :vartype status_message: str
     :ivar target_resource: The target resource.
     :vartype target_resource: ~azure.mgmt.resource.resources.v2018_02_01.models.TargetResource
     :ivar request: The HTTP request message.
@@ -340,7 +340,7 @@ class DeploymentOperationProperties(msrest.serialization.Model):
         'timestamp': {'key': 'timestamp', 'type': 'iso-8601'},
         'service_request_id': {'key': 'serviceRequestId', 'type': 'str'},
         'status_code': {'key': 'statusCode', 'type': 'str'},
-        'status_message': {'key': 'statusMessage', 'type': 'object'},
+        'status_message': {'key': 'statusMessage', 'type': 'str'},
         'target_resource': {'key': 'targetResource', 'type': 'TargetResource'},
         'request': {'key': 'request', 'type': 'HttpMessage'},
         'response': {'key': 'response', 'type': 'HttpMessage'},
@@ -399,7 +399,7 @@ class DeploymentProperties(msrest.serialization.Model):
      syntax directly in the request rather than link to an existing template. It can be a JObject or
      well-formed JSON string. Use either the templateLink property or the template property, but not
      both.
-    :type template: object
+    :type template: str
     :param template_link: The URI of the template. Use either the templateLink property or the
      template property, but not both.
     :type template_link: ~azure.mgmt.resource.resources.v2018_02_01.models.TemplateLink
@@ -407,7 +407,7 @@ class DeploymentProperties(msrest.serialization.Model):
      You use this element when you want to provide the parameter values directly in the request
      rather than link to an existing parameter file. Use either the parametersLink property or the
      parameters property, but not both. It can be a JObject or a well formed JSON string.
-    :type parameters: object
+    :type parameters: str
     :param parameters_link: The URI of parameters file. You use this element to link to an existing
      parameters file. Use either the parametersLink property or the parameters property, but not
      both.
@@ -430,9 +430,9 @@ class DeploymentProperties(msrest.serialization.Model):
     }
 
     _attribute_map = {
-        'template': {'key': 'template', 'type': 'object'},
+        'template': {'key': 'template', 'type': 'str'},
         'template_link': {'key': 'templateLink', 'type': 'TemplateLink'},
-        'parameters': {'key': 'parameters', 'type': 'object'},
+        'parameters': {'key': 'parameters', 'type': 'str'},
         'parameters_link': {'key': 'parametersLink', 'type': 'ParametersLink'},
         'mode': {'key': 'mode', 'type': 'str'},
         'debug_setting': {'key': 'debugSetting', 'type': 'DebugSetting'},
@@ -465,18 +465,18 @@ class DeploymentPropertiesExtended(msrest.serialization.Model):
     :ivar timestamp: The timestamp of the template deployment.
     :vartype timestamp: ~datetime.datetime
     :param outputs: Key/value pairs that represent deployment output.
-    :type outputs: object
+    :type outputs: str
     :param providers: The list of resource providers needed for the deployment.
     :type providers: list[~azure.mgmt.resource.resources.v2018_02_01.models.Provider]
     :param dependencies: The list of deployment dependencies.
     :type dependencies: list[~azure.mgmt.resource.resources.v2018_02_01.models.Dependency]
     :param template: The template content. Use only one of Template or TemplateLink.
-    :type template: object
+    :type template: str
     :param template_link: The URI referencing the template. Use only one of Template or
      TemplateLink.
     :type template_link: ~azure.mgmt.resource.resources.v2018_02_01.models.TemplateLink
     :param parameters: Deployment parameters. Use only one of Parameters or ParametersLink.
-    :type parameters: object
+    :type parameters: str
     :param parameters_link: The URI referencing the parameters. Use only one of Parameters or
      ParametersLink.
     :type parameters_link: ~azure.mgmt.resource.resources.v2018_02_01.models.ParametersLink
@@ -500,12 +500,12 @@ class DeploymentPropertiesExtended(msrest.serialization.Model):
         'provisioning_state': {'key': 'provisioningState', 'type': 'str'},
         'correlation_id': {'key': 'correlationId', 'type': 'str'},
         'timestamp': {'key': 'timestamp', 'type': 'iso-8601'},
-        'outputs': {'key': 'outputs', 'type': 'object'},
+        'outputs': {'key': 'outputs', 'type': 'str'},
         'providers': {'key': 'providers', 'type': '[Provider]'},
         'dependencies': {'key': 'dependencies', 'type': '[Dependency]'},
-        'template': {'key': 'template', 'type': 'object'},
+        'template': {'key': 'template', 'type': 'str'},
         'template_link': {'key': 'templateLink', 'type': 'TemplateLink'},
-        'parameters': {'key': 'parameters', 'type': 'object'},
+        'parameters': {'key': 'parameters', 'type': 'str'},
         'parameters_link': {'key': 'parametersLink', 'type': 'ParametersLink'},
         'mode': {'key': 'mode', 'type': 'str'},
         'debug_setting': {'key': 'debugSetting', 'type': 'DebugSetting'},
@@ -565,7 +565,7 @@ class ErrorAdditionalInfo(msrest.serialization.Model):
     :ivar type: The additional info type.
     :vartype type: str
     :ivar info: The additional info.
-    :vartype info: object
+    :vartype info: str
     """
 
     _validation = {
@@ -575,7 +575,7 @@ class ErrorAdditionalInfo(msrest.serialization.Model):
 
     _attribute_map = {
         'type': {'key': 'type', 'type': 'str'},
-        'info': {'key': 'info', 'type': 'object'},
+        'info': {'key': 'info', 'type': 'str'},
     }
 
     def __init__(
@@ -720,7 +720,7 @@ class GenericResource(Resource):
     :param plan: The plan of the resource.
     :type plan: ~azure.mgmt.resource.resources.v2018_02_01.models.Plan
     :param properties: The resource properties.
-    :type properties: object
+    :type properties: str
     :param kind: The kind of the resource.
     :type kind: str
     :param managed_by: ID of the resource that manages this resource.
@@ -745,7 +745,7 @@ class GenericResource(Resource):
         'location': {'key': 'location', 'type': 'str'},
         'tags': {'key': 'tags', 'type': '{str}'},
         'plan': {'key': 'plan', 'type': 'Plan'},
-        'properties': {'key': 'properties', 'type': 'object'},
+        'properties': {'key': 'properties', 'type': 'str'},
         'kind': {'key': 'kind', 'type': 'str'},
         'managed_by': {'key': 'managedBy', 'type': 'str'},
         'sku': {'key': 'sku', 'type': 'Sku'},
@@ -783,7 +783,7 @@ class GenericResourceExpanded(GenericResource):
     :param plan: The plan of the resource.
     :type plan: ~azure.mgmt.resource.resources.v2018_02_01.models.Plan
     :param properties: The resource properties.
-    :type properties: object
+    :type properties: str
     :param kind: The kind of the resource.
     :type kind: str
     :param managed_by: ID of the resource that manages this resource.
@@ -820,7 +820,7 @@ class GenericResourceExpanded(GenericResource):
         'location': {'key': 'location', 'type': 'str'},
         'tags': {'key': 'tags', 'type': '{str}'},
         'plan': {'key': 'plan', 'type': 'Plan'},
-        'properties': {'key': 'properties', 'type': 'object'},
+        'properties': {'key': 'properties', 'type': 'str'},
         'kind': {'key': 'kind', 'type': 'str'},
         'managed_by': {'key': 'managedBy', 'type': 'str'},
         'sku': {'key': 'sku', 'type': 'Sku'},
@@ -871,11 +871,11 @@ class HttpMessage(msrest.serialization.Model):
     """HTTP message.
 
     :param content: HTTP message content.
-    :type content: object
+    :type content: str
     """
 
     _attribute_map = {
-        'content': {'key': 'content', 'type': 'object'},
+        'content': {'key': 'content', 'type': 'str'},
     }
 
     def __init__(
@@ -1123,6 +1123,8 @@ class ProviderResourceType(msrest.serialization.Model):
     :type aliases: list[~azure.mgmt.resource.resources.v2018_02_01.models.AliasType]
     :param api_versions: The API version.
     :type api_versions: list[str]
+    :param zone_mappings:
+    :type zone_mappings: list[~azure.mgmt.resource.resources.v2018_02_01.models.ZoneMapping]
     :param properties: The properties.
     :type properties: dict[str, str]
     """
@@ -1132,6 +1134,7 @@ class ProviderResourceType(msrest.serialization.Model):
         'locations': {'key': 'locations', 'type': '[str]'},
         'aliases': {'key': 'aliases', 'type': '[AliasType]'},
         'api_versions': {'key': 'apiVersions', 'type': '[str]'},
+        'zone_mappings': {'key': 'zoneMappings', 'type': '[ZoneMapping]'},
         'properties': {'key': 'properties', 'type': '{str}'},
     }
 
@@ -1144,6 +1147,7 @@ class ProviderResourceType(msrest.serialization.Model):
         self.locations = kwargs.get('locations', None)
         self.aliases = kwargs.get('aliases', None)
         self.api_versions = kwargs.get('api_versions', None)
+        self.zone_mappings = kwargs.get('zone_mappings', None)
         self.properties = kwargs.get('properties', None)
 
 
@@ -1200,14 +1204,14 @@ class ResourceGroupExportResult(msrest.serialization.Model):
     """Resource group export result.
 
     :param template: The template content.
-    :type template: object
+    :type template: str
     :param error: The error.
     :type error:
      ~azure.mgmt.resource.resources.v2018_02_01.models.ResourceManagementErrorWithDetails
     """
 
     _attribute_map = {
-        'template': {'key': 'template', 'type': 'object'},
+        'template': {'key': 'template', 'type': 'str'},
         'error': {'key': 'error', 'type': 'ResourceManagementErrorWithDetails'},
     }
 
@@ -1714,3 +1718,26 @@ class TemplateLink(msrest.serialization.Model):
         super(TemplateLink, self).__init__(**kwargs)
         self.uri = kwargs['uri']
         self.content_version = kwargs.get('content_version', None)
+
+
+class ZoneMapping(msrest.serialization.Model):
+    """ZoneMapping.
+
+    :param location: The location of the zone mapping.
+    :type location: str
+    :param zones:
+    :type zones: list[str]
+    """
+
+    _attribute_map = {
+        'location': {'key': 'location', 'type': 'str'},
+        'zones': {'key': 'zones', 'type': '[str]'},
+    }
+
+    def __init__(
+        self,
+        **kwargs
+    ):
+        super(ZoneMapping, self).__init__(**kwargs)
+        self.location = kwargs.get('location', None)
+        self.zones = kwargs.get('zones', None)