+++ sdk/synapse/azure-synapse-artifacts/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Release History
 
-## 0.7.0 (Unreleased)
+## 0.8.0 (Unreleased)
+
+
+## 0.7.0 (2021-05-11)
 
 ### Bug fixes